add file load to the new scenegraph load dialog
