+++ src/components/common/LoadSceneGraphDialog.module.css
@@ -15,9 +15,10 @@
   background: white;
   border-radius: 8px;
   padding: 20px;
-  min-width: 400px;
-  max-width: 600px;
-  max-height: 80vh;
+  width: 80vw;
+  height: 80vh;
+  display: flex;
+  flex-direction: column;
 }
 
 .dark .dialog {
@@ -46,6 +47,51 @@
 
 .dark .closeButton {
   color: #999;
+}
+
+.tabs {
+  display: flex;
+  margin-bottom: 16px;
+  border-bottom: 1px solid #ddd;
+}
+
+.tabButton {
+  flex: 1;
+  padding: 8px;
+  background: none;
+  border: none;
+  cursor: pointer;
+  text-align: center;
+  font-size: 16px;
+  color: #666;
+}
+
+.tabButton:hover {
+  background: rgba(0, 123, 255, 0.1);
+}
+
+.activeTab {
+  font-weight: bold;
+  border-bottom: 2px solid #007bff;
+  color: #007bff;
+}
+
+.fileTab {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex: 1;
+}
+
+.fileInput {
+  padding: 8px;
+  font-size: 16px;
+}
+
+.demosTab {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
 }
 
 .toolbar {
@@ -93,9 +139,9 @@
 .content {
   display: flex;
   flex-direction: column;
-  gap: 4px; /* Reduced spacing between categories */
+  gap: 4px;
   overflow-y: auto;
-  max-height: 60vh;
+  flex: 1;
 }
 
 .treeNode {
@@ -111,7 +157,7 @@
   background: #f5f5f5;
   border-radius: 4px;
   cursor: pointer;
-  user-select: none; /* Prevent text selection when clicking */
+  user-select: none;
 }
 
 .treeNodeHeader.light {
@@ -123,7 +169,7 @@
 }
 
 .treeNodeHeader:hover {
-  background: rgba(0, 123, 255, 0.1); /* Highlight on hover */
+  background: rgba(0, 123, 255, 0.1);
 }
 
 .expandButton {
@@ -135,7 +181,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  pointer-events: none; /* Prevent button from intercepting clicks */
+  pointer-events: none;
 }
 
 .dark .expandButton {
